Move list to get clusters command, matching Kind
